OOP exercise incl. unit tests
OOP exercise - working with classes, methods, properties (getters and setters). The program creates rectangles, calculates their area and circumference, can increase their size and print out information into the console.

Unit tests check that the constructors work properly and won't allow an instance of a rectangle with one side smaller or equal to 0. The second set of tests similarly checks the methods that increase the size of the rectangle.